backend/requirements.txt
google-api-python-client==2.170.0
google-auth==2.40.3
